Modularise VaultThemeOnJoin
